+++ src/bl64-log.bash
@@ -35,7 +35,7 @@ function bl64_log_set_runtime() {
 #   BL64_LIB_ERROR_MODULE_SETUP_INVALID
 #######################################
 function _bl64_log_register() {
-  bl64_dbg_lib_log_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_log_is_enabled && bl64_dbg_lib_show_function "$@"
   local source="$1"
   local category="$2"
   local payload="$3"
@@ -82,7 +82,7 @@ function _bl64_log_register() {
 #   >0: failed to save the log record
 #######################################
 function bl64_log_info() {
-  bl64_dbg_lib_log_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_log_is_enabled && bl64_dbg_lib_show_function "$@"
   local source="$1"
   local payload="$2"
 
@@ -111,7 +111,7 @@ function bl64_log_info() {
 #   >0: failed to save the log record
 #######################################
 function bl64_log_error() {
-  bl64_dbg_lib_log_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_log_is_enabled && bl64_dbg_lib_show_function "$@"
   local source="$1"
   local payload="$2"
 
@@ -137,7 +137,7 @@ function bl64_log_error() {
 #   >0: failed to save the log record
 #######################################
 function bl64_log_warning() {
-  bl64_dbg_lib_log_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_log_is_enabled && bl64_dbg_lib_show_function "$@"
   local source="$1"
   local payload="$2"