+++ src/bl64-msg.bash
@@ -3,6 +3,16 @@
 #######################################
 
 #
+# Deprecation aliases
+#
+# * Aliases to deprecated functions 
+# * Needed to maintain compatibility up to N-2 versions
+#
+
+function bl64_msg_app_verbose_enabled { bl64_msg_show_deprecated 'bl64_msg_app_verbose_enabled' 'bl64_msg_app_verbose_is_enabled'; bl64_msg_app_verbose_is_enabled; }
+function bl64_msg_lib_verbose_enabled { bl64_msg_show_deprecated 'bl64_msg_lib_verbose_enabled' 'bl64_msg_lib_verbose_is_enabled'; bl64_msg_lib_verbose_is_enabled; }
+
+#
 # Internal functions
 #
 
@@ -55,7 +65,7 @@ function _bl64_msg_alert_show_parameter() {
 #   BL64_LIB_ERROR_MODULE_SETUP_INVALID
 #######################################
 function _bl64_msg_print() {
-  bl64_dbg_lib_msg_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
   local attribute="${1:-}"
   local type="${2:-}"
   local message="${3:-}"
@@ -71,7 +81,7 @@ function _bl64_msg_print() {
 }
 
 function _bl64_msg_format_ansi() {
-  bl64_dbg_lib_msg_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
   local attribute="${1:-}"
   local type="${2:-}"
   local message="${3:-}"
@@ -122,7 +132,7 @@ function _bl64_msg_format_ansi() {
 }
 
 function _bl64_msg_format_ascii() {
-  bl64_dbg_lib_msg_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
   local attribute="${1:-}"
   local type="${2:-}"
   local message="${3:-}"
@@ -193,7 +203,7 @@ function _bl64_msg_format_ascii() {
 #   >0: printf error
 #######################################
 function bl64_msg_show_usage() {
-  bl64_dbg_lib_msg_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
   local usage="${1:-${BL64_VAR_NULL}}"
   local description="${2:-${BL64_VAR_DEFAULT}}"
   local commands="${3:-${BL64_VAR_DEFAULT}}"
@@ -236,7 +246,7 @@ function bl64_msg_show_usage() {
 #   >0: printf error
 #######################################
 function bl64_msg_show_error() {
-  bl64_dbg_lib_msg_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
   local message="$1"
 
   bl64_log_error "${FUNCNAME[1]:-MAIN}" "$message" &&
@@ -256,7 +266,7 @@ function bl64_msg_show_error() {
 #   >0: printf error
 #######################################
 function bl64_msg_show_warning() {
-  bl64_dbg_lib_msg_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
   local message="$1"
 
   bl64_log_warning "${FUNCNAME[1]:-MAIN}" "$message" &&
@@ -276,11 +286,11 @@ function bl64_msg_show_warning() {
 #   >0: printf error
 #######################################
 function bl64_msg_show_init() {
-  bl64_dbg_lib_msg_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
   local message="$1"
 
   bl64_log_info "${FUNCNAME[1]:-MAIN}" "$message" &&
-    bl64_msg_app_verbose_enabled || return 0
+    bl64_msg_app_verbose_is_enabled || return 0
 
   _bl64_msg_print "$BL64_MSG_TYPE_INIT" 'Init' "$message"
 }
@@ -298,11 +308,11 @@ function bl64_msg_show_init() {
 #   >0: printf error
 #######################################
 function bl64_msg_show_info() {
-  bl64_dbg_lib_msg_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
   local message="$1"
 
   bl64_log_info "${FUNCNAME[1]:-MAIN}" "$message" &&
-    bl64_msg_app_verbose_enabled || return 0
+    bl64_msg_app_verbose_is_enabled || return 0
 
   _bl64_msg_print "$BL64_MSG_TYPE_INFO" 'Info' "$message"
 }
@@ -320,11 +330,11 @@ function bl64_msg_show_info() {
 #   >0: printf error
 #######################################
 function bl64_msg_show_phase() {
-  bl64_dbg_lib_msg_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
   local message="$1"
 
   bl64_log_info "${FUNCNAME[1]:-MAIN}" "${BL64_MSG_TYPE_PHASE}:${message}" &&
-    bl64_msg_app_verbose_enabled || return 0
+    bl64_msg_app_verbose_is_enabled || return 0
 
   _bl64_msg_print "$BL64_MSG_TYPE_PHASE" 'Phase' "${BL64_MSG_COSMETIC_PHASE_PREFIX} ${message} ${BL64_MSG_COSMETIC_PHASE_SUFIX}"
 }
@@ -342,11 +352,11 @@ function bl64_msg_show_phase() {
 #   >0: printf error
 #######################################
 function bl64_msg_show_task() {
-  bl64_dbg_lib_msg_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
   local message="$1"
 
   bl64_log_info "${FUNCNAME[1]:-MAIN}" "${BL64_MSG_TYPE_TASK}:${message}" &&
-    bl64_msg_app_verbose_enabled || return 0
+    bl64_msg_app_verbose_is_enabled || return 0
 
   _bl64_msg_print "$BL64_MSG_TYPE_TASK" 'Task' "$message"
 }
@@ -364,11 +374,11 @@ function bl64_msg_show_task() {
 #   >0: printf error
 #######################################
 function bl64_msg_show_subtask() {
-  bl64_dbg_lib_msg_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
   local message="$1"
 
   bl64_log_info "${FUNCNAME[1]:-MAIN}" "${BL64_MSG_TYPE_SUBTASK}:${message}" &&
-    bl64_msg_app_verbose_enabled || return 0
+    bl64_msg_app_verbose_is_enabled || return 0
 
   _bl64_msg_print "$BL64_MSG_TYPE_SUBTASK" 'Subtask' "${BL64_MSG_COSMETIC_ARROW2} ${message}"
 }
@@ -386,11 +396,11 @@ function bl64_msg_show_subtask() {
 #   >0: printf error
 #######################################
 function bl64_msg_show_lib_task() {
-  bl64_dbg_lib_msg_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
   local message="$1"
 
   bl64_log_info "${FUNCNAME[1]:-MAIN}" "${BL64_MSG_TYPE_LIBTASK}:${message}" &&
-    bl64_msg_lib_verbose_enabled || return 0
+    bl64_msg_lib_verbose_is_enabled || return 0
 
   _bl64_msg_print "$BL64_MSG_TYPE_LIBTASK" 'Task' "$message"
 }
@@ -408,11 +418,11 @@ function bl64_msg_show_lib_task() {
 #   >0: printf error
 #######################################
 function bl64_msg_show_lib_subtask() {
-  bl64_dbg_lib_msg_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
   local message="$1"
 
   bl64_log_info "${FUNCNAME[1]:-MAIN}" "${BL64_MSG_TYPE_LIBSUBTASK}:${message}" &&
-    bl64_msg_app_verbose_enabled || return 0
+    bl64_msg_app_verbose_is_enabled || return 0
 
   _bl64_msg_print "$BL64_MSG_TYPE_LIBSUBTASK" 'Subtask' "${BL64_MSG_COSMETIC_ARROW2} ${message}"
 }
@@ -430,11 +440,11 @@ function bl64_msg_show_lib_subtask() {
 #   >0: printf error
 #######################################
 function bl64_msg_show_lib_info() {
-  bl64_dbg_lib_msg_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
   local message="$1"
 
   bl64_log_info "${FUNCNAME[1]:-MAIN}" "${BL64_MSG_TYPE_LIBINFO}:${message}" &&
-    bl64_msg_lib_verbose_enabled || return 0
+    bl64_msg_lib_verbose_is_enabled || return 0
 
   _bl64_msg_print "$BL64_MSG_TYPE_LIBINFO" 'Info' "$message"
 }
@@ -452,11 +462,11 @@ function bl64_msg_show_lib_info() {
 #   >0: printf error
 #######################################
 function bl64_msg_show_text() {
-  bl64_dbg_lib_msg_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
   local message="$1"
 
   bl64_log_info "${FUNCNAME[1]:-MAIN}" "$message" &&
-    bl64_msg_app_verbose_enabled || return 0
+    bl64_msg_app_verbose_is_enabled || return 0
 
   printf '%s\n' "$message"
 }
@@ -476,11 +486,11 @@ function bl64_msg_show_text() {
 #   >0: printf error
 #######################################
 function bl64_msg_show_batch_start() {
-  bl64_dbg_lib_msg_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
   local message="$1"
 
   bl64_log_info "${FUNCNAME[1]:-MAIN}" "${BL64_MSG_TYPE_BATCH}:${message}" &&
-    bl64_msg_app_verbose_enabled || return 0
+    bl64_msg_app_verbose_is_enabled || return 0
 
   _bl64_msg_print "$BL64_MSG_TYPE_BATCH" 'Process' "[${message}] started"
 }
@@ -503,12 +513,12 @@ function bl64_msg_show_batch_start() {
 #######################################
 function bl64_msg_show_batch_finish() {
   local -i status=$1
-  bl64_dbg_lib_msg_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
   local message="${2-}"
 
   # shellcheck disable=SC2086
   bl64_log_info "${FUNCNAME[1]:-MAIN}" "${BL64_MSG_TYPE_BATCH}:${status}:${message}" &&
-    bl64_msg_app_verbose_enabled ||
+    bl64_msg_app_verbose_is_enabled ||
     return $status
 
   if ((status == 0)); then
@@ -535,7 +545,7 @@ function bl64_msg_show_batch_finish() {
 #   >0: printf error
 #######################################
 function bl64_msg_show_input() {
-  bl64_dbg_lib_msg_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
   local message="$1"
 
   _bl64_msg_print "$BL64_MSG_TYPE_INPUT" 'Input' "$message"
@@ -556,7 +566,7 @@ function bl64_msg_show_input() {
 #   >0: printf error
 #######################################
 function bl64_msg_show_separator() {
-  bl64_dbg_lib_msg_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
   local message="${1:-$BL64_VAR_DEFAULT}"
   local separator="${2:-$BL64_VAR_DEFAULT}"
   local length="${3:-$BL64_VAR_DEFAULT}"
@@ -592,7 +602,7 @@ function bl64_msg_show_separator() {
 #   >0: printf error
 #######################################
 function bl64_msg_show_deprecated() {
-  bl64_dbg_lib_msg_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
   local feature="${1:-}"
   local replacement="${2:-non-available}"
 
@@ -614,7 +624,7 @@ function bl64_msg_show_deprecated() {
 #   >0: printf error
 #######################################
 function bl64_msg_show_setup() {
-  bl64_dbg_lib_msg_enabled && bl64_dbg_lib_show_function "$@"
+  bl64_dbg_lib_msg_is_enabled && bl64_dbg_lib_show_function "$@"
   local message="${1:-$BL64_VAR_DEFAULT}"
   local variable=''