+++ src/bl64-rbac.bash
@@ -93,7 +93,7 @@ function bl64_rbac_check_sudoers() {
     bl64_check_command "$BL64_RBAC_CMD_VISUDO" ||
     return $?
 
-  bl64_dbg_lib_command_enabled && debug=' '
+  bl64_dbg_lib_command_is_enabled && debug=' '
 
   # shellcheck disable=SC2086
   "$BL64_RBAC_CMD_VISUDO" \